+++ runtime-transaction/src/instruction_data_len.rs
@@ -2,15 +2,15 @@ use {solana_pubkey::Pubkey, solana_svm_transaction::instruction::SVMInstruction}
 
 #[derive(Default)]
 pub struct InstructionDataLenBuilder {
-    value: u64,
+    value: u16,
 }
 
 impl InstructionDataLenBuilder {
     pub fn process_instruction(&mut self, _program_id: &Pubkey, instruction: &SVMInstruction) {
-        self.value = self.value.saturating_add(instruction.data.len() as u64);
+        self.value = self.value.saturating_add(instruction.data.len() as u16);
     }
 
-    pub fn build(self) -> u64 {
+    pub fn build(self) -> u16 {
         self.value
     }
 }
